Move whole game to main component
